### 2016-09-22 #76 @kyleniemeyer
Hi @jordigh & alex-konovalov: would either of you be able to review this submission? I've also requested a review from an external subject-area expert, so just one of you would be great.
