### 2016-09-22 #79 @whedon
[REVIEW]: Python Active-subspaces Utility Library
**Submitting author:** @paulcon (<a href="http://orcid.org/0000-0003-3726-6307">Paul Constantine</a>)
**Repository:** <a href="https://github.com/paulcon/active_subspaces" target ="_blank">https://github.com/paulcon/active_subspaces</a>
**Version:** v 0.1

- [x] **Version:** Does the release version given match the GitHub release (v 0.1)?
