### 2016-09-22 #79 @nicoguaro
I installed the software in Windows 7 and ran the tests, that were successful.

I was trying to run the Jupyter notebook in the folder ``tutorials``, but it returns the following error in cell number 4.

    <ipython-input-4-1bdb1654efc5> in <module>()
          1 #This plots the eigenvalues (ss.eigenvalues) with bootstrap ranges (ss.e_br)
    ----> 2 ac.utils.plotters.eigenvalues(ss.eigenvalues, ss.e_br)
          3
          4 #This plots subspace errors with bootstrap ranges (all contained in ss.sub_br)
          5 ac.utils.plotters.subspace_errors(ss.sub_br)

    AttributeError: Subspaces instance has no attribute 'eigenvalues'
