### 2016-09-23 #79 @mtezzele
Minor issue regarding the version: the release is 0.1 and the installed (now) is 0.1.1.
