### 2016-09-23 550d9ac
update paper.md
