### 2016-09-23 02f3709
personal fix grad
