+++ active_subspaces/gradients.py
@@ -38,9 +38,10 @@ def local_linear_gradients(X, f, p=None, weights=None):
     if weights is None:
         weights = np.ones((M, 1)) / M
 
-    MM = np.minimum(int(np.ceil(10*m*np.log(m))), M)
+    MM = np.minimum(int(np.ceil(10*m*np.log(m)))*100, M)
     logging.getLogger(__name__).debug('Computing {:d} local linear approximations with {:d} points in {:d} dims.'.format(MM, M, m))
     df = np.zeros((MM, m))
+
     for i in range(MM):
         ii = np.random.randint(M)
         x = X[ii,:]
