### 2016-09-24 3c7fd12
remove divide by zero, only Chuck Norris can do that
