### 2016-09-24 #79 @paulcon
Thanks for the help, reviewers! A few revisions:

(1) I have fixed the missing attribute errors in the notebooks. That was a consequence of a previous refactoring.
(2) I have corrected the version in setup.py to be consistent with the version release.
(3) I modified the code in utils/reponse_surfaces.py in such a way to remove the divide-by-zero RuntimeWarning.

Comments for @nicoguaro:
We have not developed/tested with Windows 10 / Anaconda, and we do not have easy access to such a system. All your failed tests fail on the same function. As best I can tell, this is not an issue with our code. I can find no bug, and tests pass on our systems. I'm not sure what to do with that one. Please advise.

Comments for @mtezzele:
(1) I have no specific guidelines for contributing. Anything that I would write would look like a github tutorial for forking, making a change, and submitting a pull request. I've also never written such a section, and I don't have any examples. If you have an example of such a section that I could mimic, please let me know.
(2) Regarding the documentation, we had mixed results using sphinx to build docs locally on different systems. It worked on some but not others. Therefore, we went with ReadTheDocs. It is set up to update automatically via a github webhook. So everyone should have access to the latest docs---based on the docstrings comments---through ReadTheDocs. To me, this is sufficient given the mixed success we had building sphinx locally.

Thanks!
