### 2016-09-24 #79 @mtezzele
@paulcon I emailed to you an example of guidelines.

Some functions are not documented. See the last ones in the domains.py file for example.

After fixing these minor issues, to me the software is well written. Tests cover 90% of the code, documentation is clear and the examples provided explain the usage of the software well.

I suggest to accept this work on JOSS.
