### 2016-09-24 #79 @arfon
@mtezzele thanks for the review here :zap: @nicoguaro - do you agree?

@paulcon if you can make these final changes (and pending @nicoguaro's approval) we can move forward with accepting this. Please note, once all changes have been made we need you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive.
