### 2016-09-25 4070f0a
adding community guidelines to readme
