### 2016-09-25 3a08391
updating documentation coverage
