### 2016-09-25 #79 @paulcon
I have made the following changes:

(1) Updated community guidelines following a helpful example from @mtezzele
(2) Expanded documentation for functions in domains.py, subpsaces.py, and utils/misc.py following the suggestions from @mtezzele. These are now updated in the latest build on ReadTheDocs.

If there anything else, please let me know. Thanks!
