### 2016-09-25 #79 @kyleniemeyer
Thanks @paulcon!

@mtezzele and @nicoguaro are your concerns/questions all satisfied? I see that a few checkboxes at the top of the issue are still empty—have those issues been resolved?
