### 2016-09-26 #79 @nicoguaro
@kyleniemeyer, I have some comments

- It works fine in Linux Mint 18 (64 bits).
- I installed Gurobi (in Linux), but I can't tell if it is being used or not. When I ran the test, the elapsed time was about the same.
- The software works under Windows, if Python 32 bits is used.

Regarding the documentation that it should include more than just the docstrings from the code. Some points that I consider pertinent are below. Some examples (http://barbagroup.github.io/pygbe/docs/, http://cvxopt.org/documentation/index.html)

- Description of the software.
- Installation.
- How to use the software, examples. There is no need to repeat the notebooks, but maybe one or two examples. Considering that some of the examples are more expensive to run than others (like Ebola). Maybe a toy example (like this video: https://youtu.be/mJvKzjT6lmY) would be good.
- Link to the repository.
- FAQ, like in the website: http://activesubspaces.org/faq/
