### 2016-09-26 #79 @paulcon
Thanks, @nicoguaro. These are excellent suggestions for the documentation. I can have them done by next week.
