### 2016-09-27 c43e929
implemented parallelization for simrunners
