### 2016-09-27 c8c2f19
altered parallelization
