+++ active_subspaces/utils/simrunners.py
@@ -1,9 +1,16 @@
 """Utilities for running several simulations at different inputs."""
 
 import numpy as np
-import multiprocessing as mp
 import time
 from misc import process_inputs
+
+# checking to see if system has gurobi
+try:
+    HAS_MP = True
+    import multiprocessing as mp
+except ImportError, e:
+    HAS_MP = False
+    pass
 
 class SimulationRunner():
     """A class for running several simulations at different input values.
@@ -42,7 +49,7 @@ class SimulationRunner():
 
         self.fun = fun
 
-    def run(self, X, parallel=True, num_cores=mp.cpu_count()-1):
+    def run(self, X, parallel=True, num_cores=None):
         """Run the simulation at several input values.
         
         Parameters
@@ -74,23 +81,25 @@ class SimulationRunner():
 
         X, M, m = process_inputs(X)
         F = np.zeros((M, 1))
+
+        # TODO: provide some timing information
+        # start = time.time()
         
-        # Use parallel computing if desired and num_cores makes sense
-        if parallel and isinstance(num_cores, int)\
+        # Set num_cores to its default value if multiprocessing is present
+        # and the user hasn't specified a value.
+        if parallel and HAS_MP and num_cores is None: num_cores = mp.cpu_count() - 1
+        
+        # Use parallel computing if desired and num_cores makes sense.
+        if parallel and HAS_MP and isinstance(num_cores, int)\
         and num_cores >= 1 and num_cores <= mp.cpu_count():
             pool = mp.Pool(processes=num_cores)
             F = np.array(pool.map(self.fun, X)).reshape((M, 1))
             pool.close()
             pool.join()
-            return F
-
-
-        # TODO: provide some timing information
-        # start = time.time()
-        
-        for i in range(M):
-            F[i] = self.fun(X[i,:].reshape((1,m)))
-            
+        else: # Otherwise use a serial loop.        
+            for i in range(M):
+                F[i] = self.fun(X[i,:].reshape((1,m)))
+                
         # TODO: provide some timing information
         # end = time.time() - start
 
@@ -138,7 +147,7 @@ class SimulationGradientRunner():
 
         self.dfun = dfun
 
-    def run(self, X, parallel=True, num_cores=mp.cpu_count()-1):
+    def run(self, X, parallel=True, num_cores=None):
         """Run at several input values.
         
         Run the simulation at several input values and return the gradients of
@@ -175,21 +184,24 @@ class SimulationGradientRunner():
         X, M, m = process_inputs(X)
         dF = np.zeros((M, m))
 
+        # TODO: provide some timing information
+        # start = time.time()
+
+        # Set num_cores to its default value if multiprocessing is present
+        # and the user hasn't specified a value.
+        if parallel and HAS_MP and num_cores is None: num_cores = mp.cpu_count() - 1
+
         # Use parallel computing if desired and num_cores makes sense
-        if parallel and isinstance(num_cores, int)\
+        if parallel and HAS_MP and isinstance(num_cores, int)\
         and num_cores >= 1 and num_cores <= mp.cpu_count():
             pool = mp.Pool(processes=num_cores)
             dF = np.array(pool.map(self.dfun, X)).squeeze()
             pool.close()
             pool.join()
-            return dF
-
-        # TODO: provide some timing information
-        # start = time.time()
-        
-        for i in range(M):
-            df = self.dfun(X[i,:].reshape((1,m)))
-            dF[i,:] = df.reshape((1,m))
+        else: # Otherwise use a serial loop.        
+            for i in range(M):
+                df = self.dfun(X[i,:].reshape((1,m)))
+                dF[i,:] = df.reshape((1,m))
         
         # TODO: provide some timing information
         # end = time.time() - start
