### 2016-09-29 8794ee3
updating documentation for JOSS
