+++ docs/source/conf.py
@@ -58,7 +58,7 @@ author = u'Paul Constantine'
 # built documents.
 #
 # The short X.Y version.
-version = '0.1.1'
+version = '0.1'
 # The full version, including alpha/beta/rc tags.
 release = version
 
