### 2016-09-29 #79 @paulcon
I have updated the documentation following @nicoguaro's helpful suggestions. The only one I did not include was the FAQ. Instead, I pointed interested readers to the Active Subspaces website and the associated SIAM book.

Please let me know if you need anything else. Thanks!
