### 2016-09-29 #79 @paulcon
Oops... documentation example has a bug. One second...
