### 2016-09-29 4d5abdb
fixing doc example
