### 2016-09-29 d00f32a
fix bug in response_surfaces that made test fail
