+++ active_subspaces/utils/response_surfaces.py
@@ -252,7 +252,7 @@ class RadialBasisApproximation(ResponseSurface):
                 v = g*np.sum(e[m:])*np.ones(f.shape)
         
         # ensure conditioning
-        v = np.amax([v, 1e-6*np.ones(f.shape)], axis=0)
+        v = np.amax([v.reshape(f.shape), 1e-6*np.ones(f.shape)], axis=0)
     
         # covariance matrix of observations
         K = exponential_squared(X, X, 1.0, ell)
