### 2016-09-29 #79 @kyleniemeyer
Thanks @paulcon! @nicoguaro it looks to me that your remaining issues have been satisfied, can you confirm?

If so, @paulcon the last step is for you to archive the latest version with Zenodo or figshare (e.g.), then let us know the resuting DOI here. (I'd also encourage you to add a badge to the README with the DOI.)
