### 2016-09-29 #79 @nicoguaro
@kyleniemeyer, @paulcon, it is good to go.

As a comment to @kyleniemeyer, can the generated PDF include the affiliation of authors 1-4 just once and use the superindex 1 to all of them?
