### 2016-09-29 2ce1575
adding DOI badge to README
