+++ README.md
@@ -1,5 +1,7 @@
 # Description
 
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.158941.svg)](https://doi.org/10.5281/zenodo.158941)
+
 [Active subspaces](http://activesubspaces.org) are part of an emerging set of tools for discovering low-dimensional structure in a given function of several variables. Interesting applications arise in deterministic computer simulations of complex physical systems, where the function is the map from the physical model's input parameters to its output quantity of interest. The active subspace is the span of particular directions in the input parameter space; perturbing the inputs along these *active* directions changes the output more, on average, than perturbing the inputs orthogonally to the active directions. By focusing on the model's response along active directions and ignoring the relatively inactive directions, we *reduce the dimension* for parameter studies---such as optimization and integration---that are essential to engineering tasks such as design and uncertainty quantification.
 
 This library contains Python tools for discovering and exploiting a given model's active subspace. The user may provide a function handle to a complex model or its gradient with respect to the input parameters. Alternatively, the user may provide a set of input/output pairs from a previously executed set of runs (e.g., a Monte Carlo or Latin hypercube study). Essential classes and methods are documented; see documentation below. We also provide a set of [Jupyter](http://jupyter.org/) notebooks that demonstrate how to apply the code to a given model.
