### 2016-09-29 b9ccb0f
adding readthedocs badge to README
