### 2016-09-29 573d35b
restructure testing in README
