### 2016-09-29 #79 @paulcon
Here's the DOI:
http://doi.org/10.5281/zenodo.158941

What is the proper way to cite this, e.g., in BibTex?
