### 2016-09-29 #79 @arfon
> As a comment to @kyleniemeyer, can the generated PDF include the affiliation of authors 1-4 just once and use the superindex 1 to all of them?

@nicoguaro - can you reformat the paper metadata to look like this?  https://raw.githubusercontent.com/msmbuilder/osprey/master/paper/paper.md
