### 2016-09-29 #79 @kyleniemeyer
@arfon that should be directed at @paulcon 😄

@paulcon That is a great question—I would normally do "Author, Title + version #, year, Zenodo, DOI". Now that you have the JOSS paper though, you can cite that instead (or in addition to).
