### 2016-09-29 #79 @arfon
> @arfon that should be directed at @paulcon 😄

Opps, sorry. @paulcon - can you reformat the paper metadata to look like this? https://raw.githubusercontent.com/msmbuilder/osprey/master/paper/paper.md
