### 2016-09-29 23149bb
updating affiliations in paper.md
