### 2016-09-29 #79 @arfon
> Done. Thanks!

Great. If this paper looks good then I think we're good to accept? [10.21105.joss.00079.pdf](https://github.com/openjournals/joss-reviews/files/501490/10.21105.joss.00079.pdf)
