### 2016-09-29 #79 @arfon
@nicoguaro @mtezzele - many thanks for your reviews here.

@paulcon your paper is now accepted into JOSS. Your DOI is http://dx.doi.org/10.21105/joss.00079 🎉 🚀 💥
